chore: remove HorseTier and move MountStat to better location
