+++ common/src/main/java/com/wynntils/core/components/Models.java
@@ -34,7 +34,6 @@ import com.wynntils.models.gambits.GambitModel;
 import com.wynntils.models.gear.GearModel;
 import com.wynntils.models.gear.SetModel;
 import com.wynntils.models.guild.GuildModel;
-import com.wynntils.models.horse.MountModel;
 import com.wynntils.models.housing.HousingModel;
 import com.wynntils.models.ingredients.IngredientModel;
 import com.wynntils.models.inventory.IngredientPouchModel;
@@ -43,6 +42,7 @@ import com.wynntils.models.items.ItemEncodingModel;
 import com.wynntils.models.items.ItemModel;
 import com.wynntils.models.lootrun.LootrunModel;
 import com.wynntils.models.marker.MarkerModel;
+import com.wynntils.models.mount.MountModel;
 import com.wynntils.models.npc.NpcModel;
 import com.wynntils.models.objectives.ObjectivesModel;
 import com.wynntils.models.players.FriendsModel;
