+++ common/src/main/java/com/wynntils/models/mount/type/MountStat.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022-2026.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.items.items.game;
+package com.wynntils.models.mount.type;
 
 import java.util.Optional;
